# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service A - Lahan & Lokasi API",
        "description": "Dokumentasi API Terintegrasi untuk Manajemen Lahan dan Slot Parkir",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:3001",
            "description": "Docker Server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Melihat daftar lokasi parkir",
                "operationId": "getLocationsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Lahan-Lokasi-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized: X-IAE-KEY header is missing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden: Invalid X-IAE-KEY value"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Menambahkan master data lahan baru",
                "operationId": "storeLocation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "type",
                                    "parking_type",
                                    "total_spots",
                                    "base_rate"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Gedung Kuliah Umum Parkir"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Telekomunikasi No. 1, Bandung"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "indoor"
                                    },
                                    "parking_type": {
                                        "type": "string",
                                        "example": "regular"
                                    },
                                    "total_spots": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "base_rate": {
                                        "type": "integer",
                                        "example": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Location master data created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Lahan-Lokasi-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The name field is required."
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Melihat detail satu lokasi",
                "operationId": "getLocationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Lahan-Lokasi-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resource with ID loc_999 not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan NIM Anda (102022400039) untuk mengakses layanan",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Locations",
            "description": "Locations"
        }
    ]
}